pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["jesustr20 <jesus.tr61@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
python-decouple = "^3.8"
brevo-python = "^1.1.2"
sib-api-v3-sdk = "^7.6.0"
django-extensions = "^3.2.3"
ptpython = "^3.0.29"
ipython = "^8.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
